requirements.txt
appdirs==1.4.4
certifi==2020.12.5
chardet==3.0.4
distlib==0.3.1
docopt==0.6.2
filelock==3.0.12
idna==2.10
numpy==1.19.4
Pillow==8.0.1
pipreqs==0.4.10
requests==2.25.0
six==1.15.0
urllib3==1.26.2
virtualenv==20.2.2
yarg==0.1.9
